Correct references to flattened fields in selection signals.
Fields are only flattened if they participate in encoding rules.
Thus, we continue to use bracket notation to access datum values in
signal expressions as we cannot be guaranteed that the flattened field
will be present. However, the field name stored as part of a
SelectionProjection will *always* use dot notation as these names are
either returned to us flattened by model.vgField or users explicitly
specify them in the input spec (in which case, dot notation is their
only recourse). Thus, when referencing the field via a selection's
top-level signal, we just directly access the flattened field name,
rather than unpacking it via bracket notation.